fix(raster): resolve format gate and code-quality findings (#3099)
Restore the switch-expression indentation that dotnet format rejected in
RasterEngineCapabilityRegistry, and clear the CodeQL code-quality findings
raised on the raster storage benchmark and execution planning code:

- force double arithmetic in the aligned-mosaic origin math
- replace exact float-zero comparisons with the repo's epsilon idiom
- hoist nullable band-count reads so Selection is not re-dereferenced
- filter explicitly with Where/FirstOrDefault instead of implicit foreach filtering
- build the range test response inline so no undisposed local remains

+++ benchmarks/Honua.Benchmarks/RasterStorage/RasterStorageProtocol.cs
@@ -121,8 +121,8 @@ internal static class RasterStorageProtocol
             {
                 scenes.Add(new RasterGrid(
                     3857,
-                    originX + (column * sceneSize * scale),
-                    originY - (row * sceneSize * scale),
+                    originX + ((double)column * sceneSize * scale),
+                    originY - ((double)row * sceneSize * scale),
                     scale,
                     -scale,
                     0,
